(sync)

+++ MiniVnet/main_server.py
@@ -74,7 +74,7 @@ def handler(sock, to_handler_queue, from_handler_queue):
                     break
                 data += get_str.decode()
 
-            from_handler_queue.put(data)
+            from_handler_queue.put(data[:-1])       # Remove the "@" at the end
 
 
             # Get data from Router part
@@ -106,13 +106,29 @@ def run_router(router, _handler_process, _to_handler_queue, _from_handler_queue)
     try:
         is_continue = True      # Whether it should continue listen to requests
 
+        '''
         while is_continue:
             route_request = from_handler_queue.get()
 
             if route_request == "End Connection":
                 break
 
+            #TODO route !
             to_handler_queue.put(route_request)
+        '''
+
+        route_request = 'car_0,Pause;car_1,Pause;car_2,Pause;car_4,Pause;car_3,Pause;car_5,Pause;car_6,Pause;car_7,Pause;car_8,Pause;car_9,Pause;car_10,Pause;car_11,NEW,5.0,002_001,2,5,182.61,4;car_12,NEW,6.0,001_001,1,5,192.79,2;car_13,NEW,6.0,001_001,1,5,176.01,3;car_14,NEW,9.0,001_002,1,5,189.79,0;car_15,NEW,6.0,002_002,0,5,192.93,1;car_16,NEW,5.0,001_002,0,6,179.97,1;car_18,NEW,5.0,002_001,2,6,179.97,4;car_17,NEW,5.0,001_002,0,5,189.45,6;car_19,NEW,9.0,001_002,3,6,175.99,0;car_20,NEW,9.0,002_001,1,6,181.55,1;car_21,NEW,8.0,002_001,2,6,193.73,6;car_22,NEW,5.0,001_002,1,7,199.77,1;car_23,NEW,7.0,001_001,1,7,180.91,3;car_24,NEW,7.0,001_001,3,8,193.49,6;car_25,NEW,9.0,002_002,1,8,182.09,5;car_26,NEW,8.0,002_001,2,8,182.87,0;car_27,NEW,8.0,001_002,0,8,177.41,4;car_28,NEW,5.0,002_001,3,9,189.05,4;car_29,NEW,8.0,001_002,1,9,184.71,2;car_30,NEW,5.0,001_002,0,9,194.41,1;car_31,NEW,8.0,002_001,3,10,177.73,1;car_32,NEW,8.0,001_001,1,10,177.73,4;car_34,NEW,6.0,002_001,0,10,190.91,0;car_35,NEW,6.0,001_001,3,10,190.91,4;car_33,NEW,9.0,001_002,1,10,193.50,5;car_36,NEW,9.0,001_001,0,10,176.61,7;car_37,NEW,6.0,001_001,2,11,193.97,5;car_38,NEW,9.0,001_001,1,10,182.17,0;car_39,NEW,9.0,002_002,1,11,185.41,5;car_40,NEW,6.0,001_001,1,11,176.97,1;car_41,NEW,9.0,002_001,1,10,193.35,7;car_43,NEW,7.0,001_002,0,11,187.15,2;car_42,NEW,5.0,001_001,2,12,176.43,6;car_44,NEW,5.0,002_001,2,11,194.71,3;car_45,NEW,5.0,001_001,3,12,197.81,6;car_46,NEW,6.0,002_001,3,13,177.51,2;car_47,NEW,5.0,002_002,0,12,197.67,7;car_48,NEW,8.0,002_001,3,13,180.99,4;car_49,NEW,6.0,001_002,3,12,196.69,1;car_50,NEW,9.0,002_002,1,13,191.31,2;car_51,NEW,7.0,003_001,3,5,181.81,0;car_52,NEW,6.0,003_002,3,5,193.99,0;car_53,NEW,5.0,000_001,1,6,192.35,1;car_54,NEW,6.0,003_002,3,7,177.53,5;car_55,NEW,7.0,002_000,2,7,176.53,0;car_56,NEW,6.0,000_001,1,7,188.71,4;car_57,NEW,9.0,002_003,0,7,185.71,1;car_58,NEW,7.0,002_000,2,7,191.07,0;car_59,NEW,8.0,003_001,3,7,197.89,0;car_60,NEW,9.0,001_000,2,7,196.89,5;car_61,NEW,9.0,000_002,1,7,196.89,5;car_62,NEW,5.0,000_001,1,8,187.07,7;'
+
+        # Parse the requests
+        route_request = route_request[:-1]  # Remove the ';' at the end
+        car_request_string_list = sroute_request.split(';')
+
+        for car_request_string in car_request_string_list:
+            car_request_info_list = car_request_string.split(',')
+
+
+        print(route_request)
+
 
     except Exception as e:
         traceback.print_exc()
@@ -126,15 +142,17 @@ if __name__ == '__main__':
     handler_process = None
     try:
         # 1. Echo and tell the size of the network
-        handler_process, to_handler_queue, from_handler_queue, grid_size, scheduling_period, routing_period_num = initial_server_handler(HOST, PORT)
+        #handler_process, to_handler_queue, from_handler_queue, grid_size, scheduling_period, routing_period_num = initial_server_handler(HOST, PORT)
 
         # 2. Initialize the router
         #router = MiniVnet(grid_size, scheduling_period, routing_period_num)
 
         # 3. Start running SUMO
-        run_router(None, handler_process, to_handler_queue, from_handler_queue)
+        #run_router(None, handler_process, to_handler_queue, from_handler_queue)
+        run_router(None, None, None, None)
 
     except Exception as e:
         traceback.print_exc()
 
-    handler_process.terminate()
+    if handler_process != None:
+        handler_process.terminate()
